fix(ts-typings) Fix ts-typings (#1810)
