+++ lib/command/definitions.js
@@ -67,7 +67,7 @@ declare class Recorder {
   errHandler(fn: CallableFunction): void
   reset(): void
   session: RecorderSession
-  add(taskName, fn?: CallableFunction, force?: boolean, retry?: boolean): Promise<any>
+  add(taskName: string, fn?: CallableFunction, force?: boolean, retry?: boolean): Promise<any>
   retry(opts: Object): Promise<any>
   catch(customErrFn: CallableFunction): Promise<any>
   catchWithoutStop(customErrFn: CallableFunction ): Promise<any>
@@ -80,7 +80,7 @@ declare class Recorder {
   scheduled(): string[]
   toString(): string
   add(hookName: string, fn: CallableFunction, force?: boolean): void
-  catch(customErrFn: CallableFunction)
+  catch(customErrFn: CallableFunction): void
 }
 
 declare class CodeceptJSEvent {
@@ -118,7 +118,7 @@ declare class CodeceptJSEvent {
     before: string,
     after: string,
   }
-  emit(event: string, param: string)
+  emit(event: string, param: string): void
   cleanDispatcher(): void
 }
 
@@ -320,11 +320,12 @@ module.exports = function (genPath, options) {
   const supports = container.support(); // return all support objects
   const exportPageObjects = [];
   const injectPageObjects = [];
+  // "I" should be always, because it contains I.say
+  injectPageObjects.push(injectSupportTemplate.replace(/{{name}}/g, String('I')));
   const callbackParams = [];
   // See #1795 and #1799, there's no obvious way to use for-in with Proxies
   Reflect.ownKeys(supports).forEach((name) => {
     if (name === 'I') {
-      injectPageObjects.push(injectSupportTemplate.replace(/{{name}}/g, String(name)));
       return;
     }
     callbackParams.push(`${String(name)}?:CodeceptJS.${String(name)}`);
