+++ test/data/sandbox/configs/definitions/codecept.inject.po.json
@@ -1,0 +1,16 @@
+{
+  "tests": "./*_test.inject.po.js",
+  "timeout": 10000,
+  "output": "./output",
+  "helpers": {
+    "FileSystem": {}
+  },
+  "include": {
+    "I": "../../support/custom_steps.js",
+    "MyPage": "../../support/my_page.js",
+    "SecondPage": "../../support/second_page.js"
+  },
+  "bootstrap": false,
+  "mocha": {},
+  "name": "sandbox"
+}
