+++ test/data/sandbox/configs/definitions/fs_test.inject.po.js
@@ -1,0 +1,11 @@
+const { I, MyPage } = inject();
+
+Feature('Filesystem');
+
+Scenario('check current dir', () => {
+  console.log('injected', I, MyPage);
+  I.openDir('aaa');
+  I.seeFile('codecept.json');
+  MyPage.hasFile('uu');
+  I.seeFile('codecept.po.json');
+});
